Create 'todo.js' file for the new functionality
