+++ src/todo.js
@@ -1,0 +1,59 @@
+import Task from './task';
+
+export default class {
+  constructor() {
+    this.tasks = localStorage.tasks ? JSON.parse(localStorage.tasks) : [];
+  }
+
+  addTask(description = '') {
+    const newEntry = document.getElementById('new-item');
+    if (description || newEntry.value) {
+      if (!description) description = newEntry.value;
+      newEntry.value = '';
+      this.tasks.push(new Task(description, this.tasks.length));
+      this.updateLocalStorage();
+      this.renderList();
+    }
+  }
+
+  updateLocalStorage() {
+    localStorage.tasks = JSON.stringify(this.tasks);
+  }
+
+  renderList() {
+    const obj = this;
+
+    function toggleCompleted(e) {
+      const index = e.target.parentElement.id.split('-')[1];
+      obj.tasks[index].completed = !obj.tasks[index].completed;
+      document.getElementById(`task-${index}`).children[1].classList.toggle('completed');
+      obj.updateLocalStorage();
+      // console.table(JSON.parse(localStorage.tasks));
+    }
+
+    const list = document.getElementById('my-list');
+    list.innerHTML = '';
+    this.tasks.forEach((task) => {
+      list.innerHTML += `
+        <li id="task-${task.index}">
+          <input type="checkbox" class="checkbox" checked="${task.completed}">
+          <p class="description ${task.completed ? 'completed' : ''}">${task.description}</p>
+          <button class="li-btn bin hidden"></button>
+          <button class="li-btn v-dots"></button>
+        </li>
+      `;
+    });
+    if (this.tasks) {
+      list.innerHTML += `
+        <li>
+          <a class="clear-completed">Clear all completed</a>
+        </li>
+      `;
+    }
+    this.tasks.forEach((task) => {
+      const checkbox = document.getElementById(`task-${task.index}`).children[0];
+      checkbox.addEventListener('change', toggleCompleted);
+      checkbox.checked = task.completed;
+    });
+  }
+}
